cleaned up the user section

+++ resources/views/users/table.blade.php
@@ -5,7 +5,9 @@
                 <th>Name</th>
                 <th>Role</th>
                 <th>Email</th>
-                <th colspan="3">Action</th>
+                @if(Auth::user()->role_id < 3)
+                    <th colspan="3">Action</th>
+                @endif
             </tr>
         </thead>
         <tbody>
@@ -18,15 +20,17 @@
                 </td>
                 <td>{!! $user->role['name'] !!}</td>
                 <td>{!! $user->email !!}</td>
-                <td>
-                    {!! Form::open(['route' => ['users.destroy', $user->id], 'method' => 'delete']) !!}
-                    <div class='btn-group'>
-                        <a href="{!! route('users.show', [$user->id]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-eye-open"></i></a>
-                        <a href="{!! route('users.edit', [$user->id]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-edit"></i></a>
-                        {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
-                    </div>
-                    {!! Form::close() !!}
-                </td>
+                @if(Auth::user()->role_id < 3)
+                    <td>
+                        {!! Form::open(['route' => ['users.destroy', $user->id], 'method' => 'delete']) !!}
+                        <div class='btn-group'>
+                            <a href="{!! route('users.show', [$user->id]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-eye-open"></i></a>
+                            <a href="{!! route('users.edit', [$user->id]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-edit"></i></a>
+                            {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                        </div>
+                        {!! Form::close() !!}
+                    </td>
+                @endif
             </tr>
         @endforeach
         </tbody>
